refactor workbooks API
- extract separate modules based on functionality
- extract common utilities and models
- change test structure accordingly
Bump web-backend version: 0.8.0 → 0.8.1-dev1
