+++ web/backend/pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "uv_build"
 
 [project]
 name = "mapwisefox-web-backend"
-version = "0.8.0"
+version = "0.8.1-dev1"
 dependencies = [
     "arrow>=1.3.0",
     "authlib>=1.6.0",
@@ -32,7 +32,7 @@ module-name = "mapwisefox.web"
 namespace = true
 
 [tool.bumpversion]
-current_version = "0.8.0"
+current_version = "0.8.1-dev1"
 parse = """(?x)
     (?P<major>0|[1-9]\\d*)\\.
     (?P<minor>0|[1-9]\\d*)\\.
