+++ web/backend/src/mapwisefox/web/api/__init__.py
@@ -1,6 +1,6 @@
 from ._auth import router as auth_api_router
 from ._config import router as config_api_router
-from ._workbooks import router as workbooks_api_router
+from .workbooks import router as workbooks_api_router
 
 
 __all__ = ["auth_api_router", "config_api_router", "workbooks_api_router"]
